Delete redundant code of ztm
