+++ aries/Cargo.toml
@@ -19,7 +19,6 @@ clap = { workspace = true, features = ["derive"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing-appender = { workspace = true }
-vault = { workspace = true }
 axum = { workspace = true }
 serde_json = { workspace = true }
 serde = { workspace = true }
@@ -30,5 +29,5 @@ tower-http = { workspace = true, features = [
     "trace",
     "decompression-full",
 ] }
-ctrlc = "3.4.4"
-regex = "1.10.4"
+ctrlc = { workspace = true }
+regex = { workspace = true }
