+++ gateway/src/api/mod.rs
@@ -5,7 +5,6 @@ pub mod github_router;
 mod model;
 pub mod nostr_router;
 pub mod p2p_router;
-pub mod ztm_router;
 
 #[derive(Clone)]
 pub struct MegaApiServiceState {
