+++ gemini/Cargo.toml
@@ -26,9 +26,6 @@ tracing = { workspace = true }
 tokio = { workspace = true, features = ["net"] }
 chrono = { workspace = true }
 secp256k1 = { workspace = true, features = ["serde", "rand", "hashes"] }
-ring = "0.17.8"
-hex = { workspace = true }
-async-trait = { workspace = true }
 quinn = "0.11"
 anyhow = { workspace = true }
 lazy_static = { workspace = true }
@@ -37,6 +34,4 @@ dashmap = "6.0.1"
 uuid = { workspace = true }
 futures-util = { workspace = true }
 tokio-util = "0.7.13"
-tokio-stream = { workspace = true }
 futures = { workspace = true }
-tracing-subscriber = "0.3.19"
